Renamed notification_service to notification_provider
