+++ lib/main.dart
@@ -11,7 +11,7 @@ import 'package:nextcloud_cookbook_flutter/src/screens/loading_screen.dart';
 import './src/screens/login_screen.dart';
 import './src/screens/splash_screen.dart';
 import './src/services/user_repository.dart';
-import './src/util/notification_service.dart';
+import './src/services/notification_provider.dart';
 import 'src/blocs/authentication/authentication.dart';
 import 'src/blocs/simple_bloc_delegatae.dart';
